+++ HomeLoanApp/FormViews/Editing/PersonalDetailsEditing.swift
@@ -392,7 +392,7 @@ struct PersonalDetailsEditing: View {
     }
     
     // MARK: - identityText
-    private func identityText(location: Double) -> Double {
+    private func identityText(location: Int) -> Int {
         var location = location
         
         if identityType == 1 || identityType == 5 {
